refactor: Standardize module directory structure (views/ consistency)
Rename components/ to views/ across all modules to follow AI_DEV_PROT v15 standard

**Changes**:
- dashboard: components/ → views/ (CategoryChart, TrendChart)
- debug: components/ → views/ (ConfirmDialog)
- transaction: components/ → views/ (ImageLightbox, Pagination)
- Updated all imports to reference new paths
- Added TransactionView export to transaction/views/index.ts

**Affected Modules**:
- 02_modules/dashboard
- 02_modules/debug
- 02_modules/transaction

**Testing**:
- ✅ Build succeeds with no new errors
- ✅ Tests pass (605/611 passing - failures unrelated)
- ✅ No components/ directories remain
- ✅ All modules follow views/ convention

**Architecture Compliance**:
- Follows Pillar I (Firewalls) - consistent module structure
- Follows AI_DEV_PROT v15 standard architecture
- Matches .prot/ guidelines for module organization

Closes #190
