+++ app/src/02_modules/dashboard/views/CategoryChart.tsx
