+++ app/src/02_modules/dashboard/views/TrendChart.tsx
