--- app/src/02_modules/transaction/components/index.ts
@@ -1,2 +1,0 @@
-export { ImageLightbox } from './ImageLightbox';
-export { Pagination } from './Pagination';
