+++ app/src/02_modules/transaction/views/ImageLightbox.tsx
