+++ app/src/02_modules/transaction/views/index.ts
@@ -1,1 +1,3 @@
-export * from './TransactionView';
+export { TransactionView } from './TransactionView';
+export { ImageLightbox } from './ImageLightbox';
+export { Pagination } from './Pagination';
